autonomous_trader/requirements.txt
yfinance>=0.2.36
alpaca-trade-api>=3.2.0
pandas>=2.0.0
PyYAML>=6.0
requests>=2.31.0
pytz>=2024.1
python-dotenv>=1.0.0
schedule>=1.2.0
